+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # edge-currency-bitcoin
 
+## 2.21.7
+
+* Fix headers for bip70 payment request
+
+## 2.21.6
+
+* update the lock file to get the new bcoin with bip70
+
 ## 2.21.3
 
 * Re-enable support for Bip70